Move TRANSFORMS from __init__.py to transform_runner.py
To populate TRANSFORMS, the transform classes were being imported,
which dragged in dependencies prematurely.

This change allows pedsnetdcc to be pip installed.

Moving all settings to a new settings.py file would be a good idea,
but I'm not doing that now.

Note that indexes.py was changed slightly more than superficially
needed, which was done to avoid a circular import.
